that's a long return actually

+++ src/for_p_thread.c
@@ -41,9 +41,9 @@ void for_p_thread_destroy_mutex(pthread_mutex_t *mutex)
 
 //* Getting the number of available threads.
 //* You can thank tavianator: https://www.reddit.com/r/C_Programming/comments/6zxnr1/comment/dmzuwt6
-int32_t for_p_thread_get_cpu_threads(bool leave_room_for_main)
+size_t for_p_thread_get_cpu_threads(bool leave_room_for_main)
 {
-  int thread_count = sysconf(_SC_NPROCESSORS_ONLN);
+  size_t thread_count = (size_t)sysconf(_SC_NPROCESSORS_ONLN);
   if (leave_room_for_main)
   {
     thread_count = thread_count - 1;
